requirements.txt
django>=1.8.0
djangorestframework==3.2.4
wheel==0.24.0
# Additional requirements go here
